UPDATE
- JAVA: support upload and replace
- JS: fileroot handling?
